Backtester migrated from the private repo

+++ common/Common.Interfaces.Api/StaticData/ContractsFilter.cs
@@ -4,6 +4,6 @@ public class ContractsFilter : PagingFilter
 {
     public virtual string? Ticker { get; set; }
     public virtual long? ExchangeId { get; set; } = null;
-    public virtual List<long>? ContractIds { get; set; } = null;
+    public virtual List<int>? ContractIds { get; set; } = null;
     public virtual long? CommissionId { get; set; }
 }
